pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-harvester"
version = "0.1.0"
description = "OSINT Harvester: unified OSINT collection toolkit (DNS, WHOIS, IP reputation, HTTP headers) with export to JSON/STIX."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="Kubenew"}]
keywords = ["osint", "security", "dns", "whois", "threat-intelligence", "stix", "recon"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Topic :: Security",
]

dependencies = [
  "requests>=2.31.0",
  "dnspython>=2.6.1",
  "python-whois>=0.9.5",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "build>=1.0.0", "twine>=4.0.0"]

[project.scripts]
osint-harvester = "osint_harvester.cli:main"

[project.urls]
Homepage = "https://github.com/Kubenew/osint-harvester"
Repository = "https://github.com/Kubenew/osint-harvester"

[tool.setuptools.packages.find]
where = ["."]
include = ["osint_harvester*"]
